% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{docxToXlsxWizard}
\alias{docxToXlsxWizard}
\title{Automatically transform all tables in Word file to Excel file}
\usage{
docxToXlsxWizard(docx.file)
}
\arguments{
\item{word.docx}{Word DOCX file name (without file extensions '.docx')}
}
\description{
Parse formatted tables from Word DOCX files and move them to an Excel file.
}
\examples{
\dontrun{
  docxToXlsxWizard("Simple table")
}
}

