% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseDocxTable}
\alias{parseDocxTable}
\title{Parse formatted tables from Word DOCX files}
\usage{
parseDocxTable(word.docx)
}
\arguments{
\item{word.docx}{Word DOCX file name (with file extensions '.docx')}
}
\description{
Parse formatted tables from Word DOCX files. Note that only DOCX files are
allowed (DOC is not supported).
Also note that the table should not contain any merged or otherwise manipulated
cells in it. The function will fail to parse the rows and columns correctly.
}
\examples{
\dontrun{
  all.tables <- parseDocxTable("Demographic Tables.docx")
}
}

