% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseDocxTableComplex}
\alias{parseDocxTableComplex}
\title{Parse formatted (more complex) tables from Word DOCX files}
\usage{
parseDocxTableComplex(word.docx)
}
\arguments{
\item{word.docx}{Word DOCX file name (with file extensions '.docx')}
}
\description{
Parse formatted tables from Word DOCX files. Note that only DOCX files are
allowed (DOC is not supported).
Tables can have merged cells but pay attention that the output matches the
input table.
}
\examples{
\dontrun{
  all.tables <- parseDocxTableComplex("Demographic Tables.docx")
}
}

