% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeXlsxTables}
\alias{writeXlsxTables}
\title{Parse formatted (more complex) tables from Word DOCX files}
\usage{
writeXlsxTables(data.list, output.file.name)
}
\arguments{
\item{data.list}{Individual tables as data.frames inside a list (list elements are all data.frames)}

\item{output.file.name}{Word XLSX file name (without file extensions '.xlsx')}
}
\description{
Parse formatted tables from Word DOCX files. Note that only DOCX files are
allowed (DOC is not supported).
Tables can have merged cells but pay attention that the output matches the
input table.
}
\examples{
\dontrun{
  writeXlsxTables(demo.tables, "Demographic tables")
}
}

